WIP: Fix Structure view not being updated when changing a Rust file
When editing a Rust file, the Structure view would never be updated
until switching the Editor to a different file and back.
This happens because the Structure view never detected a change
to the PSI.

Fix this by implementing a PsiTreeChangePreprocessorBase for Rust files.

PsiManagerImpl.fireEvent notifies all psi change preprocessors
about a change to the Psi.
None of those other would accepted changes to a rust file
and execute PsiTreeCHangePreprocessorBase.onTreeChanged().

This would cause PsiModificationTrackerImpl.incOutOfCodeBlockModificationCounter()
to never be called.

PsiModificationTrackerImpl.getOutOfCodeBlockModificationCount() would always
return the same value when called by StructureTreeBuilder.MyPsiTreeChangeListener

And this would cause StructureTreeBuilder.setupUpdateAlarm to never be called
and the Structure tree would never be rebuilt.

TODO:
* This assumes all changes are "out of code block"
  which possibly pessimizes code structure updating
* In which package should the class be placed?
* In which place should the entry in plugin.xml be placed?