RES: Report an "Unresolved reference" on all paths that can not be resolved
Currently the "Unresolved reference" inspection is only active
on mod and enum paths.
For further development it is a good idea to know which references
can be resolved and which ones still need work.

Since the inspection is not enabled by default, this change will not
cause new errors for users.

TODO:
- Add tests