Use #[lang = "copy"] attribute to search for the "Copy" trait
Currently we find the "Copy" trait by looking up the derivable traits list.

This is causes problems during tests as
RsNamedElementIndex.findDerivableTraits has a check that make sure
it found the "Copy" trait inside the stdlib package.
This check makes faking the "Copy" trait in tests difficult.

Instead check the "lang" attribute as we already do for the "Sized" trait.